+++ CHANGES.md
@@ -18,7 +18,8 @@
 
 #### RPC
 
-  + ocamlformat-rpc-lib is now functorized over the IO (#1975, @gpetiot)
+  + ocamlformat-rpc-lib is now functorized over the IO (#1975, @gpetiot).
+    Now handles `Csexp.t` types instead of `Sexplib0.Sexp.t`.
 
 ### 0.20.1 (2021-12-13)